pyproject.toml
[tool.poetry]
name = "clip-backend"
version = "0.1.0"
description = ""
authors = ["NOne"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-environ = "^0.11.2"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
flex = "^6.14.1"
django-cors-headers = "^4.6.0"
channels = "^4.2.0"
channels-redis = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
